Start Screen changed to Home, Splash Screen background changed to hubWhite
