+++ app/src/main/java/com/example/myapplication/Views/ClubsView/ClubsViewModel.kt
@@ -2,6 +2,8 @@ package com.example.myapplication.Views.ClubsView
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.myapplication.DataLayer.Models.ClubModel
+import com.example.myapplication.DataLayer.Models.ClubTab
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -10,20 +12,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
 
-data class ClubModel(
-    val clubId: String = "",
-    val clubName: String = "",
-    val clubIcon: String = "",
-    val description: String = "",
-    val memberCount: Int = 0,
-    val members: List<String> = emptyList(),
-    val chatId: String = ""  // Chat ID'sini ekledik
-)
 
-enum class ClubTab {
-    MY_CLUBS,
-    ALL_CLUBS
-}
 
 class ClubsViewModel : ViewModel() {
     private val db = FirebaseFirestore.getInstance()
